LayoutTests/fast/dom/HTMLProgressElement/progress-element-form-expected.txt

CONSOLE MESSAGE: line 16: TypeError: 'undefined' is not an object (evaluating 'document.getElementById('progress1').form.id')
This test passes if the progress element can find its form.
FAIL
